refactor(mcp): extract applyDayUpdate from programPersistence to keep types-pure (#280)
Move the SupabaseClient-touching helpers (`applyDayUpdate`,
`parsedExerciseToGeneratedForApply`) out of `programPersistence.ts` and into a
dedicated `applyDayUpdate.ts`. The Deno CI typecheck on `lib/*_test.ts` walks
every transitive type import; reaching `SupabaseClient` from `programPersistence.ts`
or `format.ts` was dragging in `@types/node` and breaking the Node-less Deno
runner. Reverts the `--no-check=remote` workaround now that the chain is clean.
