+++ .github/workflows/ci.yml
@@ -52,11 +52,7 @@ jobs:
       - uses: denoland/setup-deno@v2
         with:
           deno-version: v2.x
-      # `--no-check=remote` skips typechecking of remote modules (e.g. supabase-js),
-      # which transitively reference @types/node and would otherwise require a Node
-      # toolchain in the Deno runner. Local sources are still typechecked, so the
-      # web/Edge parity guard for `programPersistence.ts` and `format.ts` holds.
-      - run: deno test --no-check=remote --allow-env "supabase/functions/mcp/lib/*_test.ts"
+      - run: deno test "supabase/functions/mcp/lib/*_test.ts" --allow-env
 
   e2e:
     runs-on: ubuntu-latest
